Enable codegen artifacts for SQL connectors
- Promote generateArtifact/generateObjectClassArtifact from
  RestBackend to ConnectorDevelopmentBackend (generic
  codegen-service client, no REST/SCIM-specific logic)
- SqlBackend no longer stubs them out, so it inherits the same
  implementation as RestBackend/ScimBackend

Task: 11926
